Move kafkaclient.TestMessageFromConsumer to streamclient
This function has no ties to the kafkaclient implementation.

The references were already updated in an earlier commit, the actual move
was left out by accident.
