+++ streamclient/kafkaclient/testing.go
@@ -41,24 +41,6 @@ func TestProducer(tb testing.TB, topic string, options ...func(c *streamconfig.P
 	return producer, func() { require.NoError(tb, producer.Close()) }
 }
 
-// TestMessageFromConsumer returns a single message, consumed from the provided
-// consumer. It has a built-in timeout mechanism to prevent the test from
-// getting stuck.
-func TestMessageFromConsumer(tb testing.TB, consumer stream.Consumer) streammsg.Message {
-	tb.Helper()
-
-	select {
-	case m := <-consumer.Messages():
-		require.NotNil(tb, m)
-
-		return m
-	case <-time.After(time.Duration(5*TestTimeoutMultiplier) * time.Second):
-		require.Fail(tb, "Timeout while waiting for message to be returned.")
-	}
-
-	return streammsg.Message{}
-}
-
 // TestMessageFromTopic returns a single message, consumed from the provided
 // topic. It has a built-in timeout mechanism to prevent the test from getting
 // stuck.
